Simplify assets uploading
Ref https://github.com/webstudio-is/webstudio-builder/issues/1335

- with persistent fetcher there is no need in assets provider
- replaced async FileReader api with sync URL.createObjectURL
  for preview while uploading

+++ packages/asset-uploader/src/patch.ts
@@ -35,5 +35,7 @@ export const patchAssets = async (
       deletedAssetIds.push(assetId);
     }
   }
-  deleteAssets({ projectId, ids: deletedAssetIds }, context);
+  if (deletedAssetIds.length !== 0) {
+    deleteAssets({ projectId, ids: deletedAssetIds }, context);
+  }
 };
